+++ src/components/Camera/CameraTypes.js
@@ -12,23 +12,24 @@ type CameraCommonType = {
   onFailure: Function,
   onUserMedia: void => void,
   i18n: Object,
-  isFullScreen: boolean
+  isFullScreen: boolean,
 }
 
 type CameraActionType = {
   handleClick: Function,
   btnText: string,
   isFullScreen: boolean,
-  btnClass: string
+  btnClass: string,
+  btnDisabled: boolean,
 }
 
 type CameraPureType = {
   ...CameraCommonType,
   hasError?: boolean,
-  onFallbackClick?: void => void,
   webcamRef: React.Ref<typeof Webcam>,
+  trackScreen: Function,
   useFullScreen: boolean => void,
-  video?: boolean
+  video?: boolean,
 }
 
 type CameraType = {
@@ -37,6 +38,7 @@ type CameraType = {
   onScreenshot: Function,
   onVideoRecorded: ?Blob => void,
   trackScreen: Function,
+  hasError?: boolean,
 }
 
 type CameraStateType = {